chore: drop jQuery dependency and split Sidebar into focused modules
jQuery was only used for three trivial classList operations on the
dアニメストア player DOM. Replace them with native classList APIs and
remove the `jquery` / `@types/jquery` dependencies (smaller bundle, one
fewer runtime).

- PlayerControllerDom: replace `$(...).removeClass/addClass` with
  classList; dedupe the identical prev/prev_thumbnail bodies into a
  single `showPrevThumbnail()` helper (behaviour unchanged).
- content/party/index.ts: replace `$(...).hasClass("show")` with
  `classList.contains`.
- Split the 630-line Sidebar.tsx into a `sidebar/` folder
  (badges, panels, SharePanel, HistoryPanel); Sidebar.tsx now only
  holds the shell (Sidebar + Panel). No behaviour or markup changes.

typecheck / lint / rspack build all pass.
